feat: replace create-react-app with degit
