+++ packages/contentful--create-contentful-app/package.json
@@ -13,7 +13,8 @@
     "node": ">=12",
     "npm": ">=6"
   },
-  "main": "lib/index.js",
+  "type": "module",
+  "main": "lib/index.mjs",
   "lint-staged": {
     "*.js": [
       "eslint --fix",
@@ -22,7 +23,7 @@
     ]
   },
   "bin": {
-    "create-contentful-app": "lib/index.js"
+    "create-contentful-app": "lib/index.mjs"
   },
   "scripts": {
     "prettier": "prettier **/*.js --write --ignore-path .gitignore",
@@ -36,10 +37,8 @@
     "@contentful/cra-template-create-contentful-app": "file:../contentful--cra-template-create-contentful-app",
     "chalk": "^4.1.0",
     "contentful-management": "^7.5.1",
-    "create-react-app": "^5.0.0",
-    "cross-spawn": "^7.0.3",
+    "degit": "2.8.4",
     "inquirer": "^8.0.0",
-    "open": "^8.0.1",
     "tildify": "^2.0.0"
   },
   "bugs": {
@@ -55,5 +54,8 @@
   ],
   "publishConfig": {
     "access": "public"
+  },
+  "devDependencies": {
+    "@types/degit": "2.8.3"
   }
 }
